lms_backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
PyPDF2==3.0.1
pyttsx3==2.99
openai==1.108.1
python-dotenv==1.1.1
Pillow==11.3.0
google-generativeai==0.3.2
requests
